Refactored CassandraSession into an interface/impl to allow simple proxying of the service inside OSGi Blueprint contexts. Fixes bug NEWTS-81.
