+++ cassandra/storage/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -22,7 +22,7 @@
     </cm:default-properties>
   </cm:property-placeholder>
 
-  <reference id="cassandraSession" ext:proxy-method="classes" interface="org.opennms.newts.cassandra.CassandraSession" />
+  <reference id="cassandraSession" interface="org.opennms.newts.cassandra.CassandraSession" />
 
   <bean id="metricRegistry" class="com.codahale.metrics.MetricRegistry" />
