+++ examples/stress/src/main/java/org/opennms/newts/stress/SelectDispatcher.java
@@ -26,6 +26,7 @@ import org.opennms.newts.api.Timestamp;
 import org.opennms.newts.api.query.ResultDescriptor;
 import org.opennms.newts.api.query.StandardAggregationFunctions;
 import org.opennms.newts.cassandra.CassandraSession;
+import org.opennms.newts.cassandra.CassandraSessionImpl;
 import org.opennms.newts.cassandra.ContextConfigurations;
 import org.opennms.newts.persistence.cassandra.CassandraSampleRepository;
 
@@ -49,7 +50,7 @@ public class SelectDispatcher extends Dispatcher {
 
         m_config = config;
 
-        CassandraSession session = new CassandraSession(
+        CassandraSession session = new CassandraSessionImpl(
                 config.getCassandraKeyspace(),
                 config.getCassandraHost(),
                 config.getCassandraPort(),
